moved NNF transformation to PreproPass
